refactor: remove useless code/badly wrote code

+++ include/memory_mapping.h
@@ -3,8 +3,8 @@
 #include <string>
 
 typedef struct {
-    unsigned long beg_addr;
-    unsigned long end_addr;
+    uintptr_t beg_addr;
+    uintptr_t end_addr;
 } mem_mapping_t;
 
 mem_mapping_t dump_mem(int, std::string);
